refactor(p1am): split poll_runtime and test_data_capture under budget (#4503)
