+++ scripts/monolith_baseline.txt
@@ -271,7 +271,7 @@ src/p1am_control_system/backend/temperature_integration.py
 src/p1am_control_system/backend/temperature_controller.py
 
 # --- Added by the P1AM SCADA safety consolidation (#4448) ---
-# These seven files crossed the 500-LOC budget by consolidating 11 safety PRs
+# These files crossed the 500-LOC budget by consolidating 11 safety PRs
 # whose fixes each fit the budget individually. `check_file_size_budget.py` did
 # not exist when that work was authored (the gate postdates the PRs, which show
 # no file-size-budget check at all), so this is inherited size, not a new
@@ -285,16 +285,10 @@ src/p1am_control_system/backend/temperature_controller.py
 # (backend/tests/test_temperature_controller.py) and the production siblings
 # data_capture.py and data_explorer_service.py.
 #
-# The genuine modularity debt is tracked, not accepted silently: poll_runtime.py
-# (216 -> 805 LOC) and test_data_capture.py (495 -> 905) both warrant splitting.
-# poll_runtime.py has clean seams (HistorianWriter, ThrottledHistorianSink,
-# HistorianRecord, _WriterCounters) that extract into a sibling module, but the
-# historian cluster is ~509 LOC on its own, so a correct split needs two modules
-# rather than one move.
-src/p1am_control_system/backend/poll_runtime.py
+# Modularity debt in poll_runtime.py and test_data_capture.py resolved in #4503.
 src/p1am_control_system/backend/pid_tuning.py
-src/p1am_control_system/backend/tests/test_data_capture.py
 src/p1am_control_system/backend/tests/test_data_explorer_service.py
 src/p1am_control_system/backend/tests/test_backend.py
 src/p1am_control_system/desktop/main_window.py
 src/p1am_control_system/calibration/calibrate.py
+
